Move reflection extensions to Reflection namespace (#7)
